requirements.txt
# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# AI/ML libraries
ultralytics==8.0.196
torch>=1.9.0
torchvision>=0.10.0
opencv-python==4.8.1.78
Pillow==10.1.0

# VietOCR
vietocr==0.3.6

# Data validation
pydantic==2.5.0

# Utilities
numpy>=1.21.0
python-dateutil==2.8.2

# Optional: for production
gunicorn==21.2.0
